pyproject.toml
[project]
name = "progetto-prova"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cairosvg>=2.7.1",
    "matplotlib>=3.9.4",
    "mplbasketball>=1.0.0",
    "nba-api>=1.6.1",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "polars>=1.17.1",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.2",
    "streamlit>=1.41.1",
]
